+++ public/scripts/power-user.js
@@ -326,6 +326,8 @@ export const power_user = {
                 right: AUTOCOMPLETE_WIDTH.CHAT,
             },
             select: AUTOCOMPLETE_SELECT_KEY.TAB + AUTOCOMPLETE_SELECT_KEY.ENTER,
+            /** Whether to show macro autocomplete in all macro-enabled fields (not just expanded editors) */
+            showInAllMacroFields: false,
         },
         parser: {
             /**@type {Object.<PARSER_FLAG,boolean>} */
@@ -1590,6 +1592,9 @@ export async function loadPowerUserSettings(settings, data) {
             if (power_user.stscript.autocomplete.select === undefined) {
                 power_user.stscript.autocomplete.select = defaultStscript.autocomplete.select;
             }
+            if (power_user.stscript.autocomplete.showInAllMacroFields === undefined) {
+                power_user.stscript.autocomplete.showInAllMacroFields = defaultStscript.autocomplete.showInAllMacroFields;
+            }
         }
         if (power_user.stscript.parser === undefined) {
             power_user.stscript.parser = defaultStscript.parser;
@@ -1727,6 +1732,7 @@ export async function loadPowerUserSettings(settings, data) {
 
     $('#stscript_autocomplete_state').val(power_user.stscript.autocomplete.state).trigger('input');
     $('#stscript_autocomplete_autoHide').prop('checked', power_user.stscript.autocomplete.autoHide ?? false).trigger('input');
+    $('#stscript_autocomplete_showInAllMacroFields').prop('checked', power_user.stscript.autocomplete.showInAllMacroFields ?? false).trigger('input');
     $('#stscript_matching').val(power_user.stscript.matching ?? 'fuzzy');
     $('#stscript_autocomplete_style').val(power_user.stscript.autocomplete.style ?? 'theme');
     document.body.setAttribute('data-stscript-style', power_user.stscript.autocomplete.style);
@@ -4076,6 +4082,11 @@ jQuery(() => {
 
     $('#stscript_autocomplete_autoHide').on('input', function () {
         power_user.stscript.autocomplete.autoHide = !!$(this).prop('checked');
+        saveSettingsDebounced();
+    });
+
+    $('#stscript_autocomplete_showInAllMacroFields').on('input', function () {
+        power_user.stscript.autocomplete.showInAllMacroFields = !!$(this).prop('checked');
         saveSettingsDebounced();
     });
